requirements.txt
# Slack Bot Dependencies
# Versiones específicas para máxima compatibilidad

# Core Slack libraries
slack-bolt>=1.18.0
slack-sdk>=3.22.0

# Required for Socket Mode and async HTTP
aiohttp>=3.8.0

# Environment variable management
python-dotenv>=1.0.0

# AI/LLM Support
anthropic>=0.18.0      # Para Claude de Anthropic
openai>=1.12.0         # Para OpenAI (también funciona con OpenRouter)
httpx>=0.24.0          # Cliente HTTP moderno para APIs

# Optional but recommended for production
# logging y monitoring
structlog>=23.1.0

# Development dependencies (opcional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0